refactor(models): extract schedule_external_review to its own module
The module-health ratchet (CI module-health-gate, base..head diff) tripped on
ticket.py: the plan_gate merge plus this branch's normalize_phase import grew
the already-over-cap file (500 cap; grandfathered at 870) to 873, and over-cap
files may only shrink. schedule_external_review is a self-contained module-level
function whose own docstring notes it lives outside Ticket for the method-count
cap — it moves to src/teatree/core/models/ticket_external_review.py and its
import sites are updated (no back-compat re-export). ticket.py drops to 846,
back under its baseline.

The new module is never imported during Django model registration, so its
sibling-model imports (Session/Task/Ticket) are top-level, not the
function-scoped imports ticket.py needs to dodge the registration-time cycle —
so no new lint suppressions, and the intra-core deferred-import ledger records
ticket.py shedding its two lazy imports. Behavior is identical.
