+++ tests/quality/deferred_import_pegs.toml
@@ -63,7 +63,7 @@
 "src/teatree/core/models/auto_review_dispatch.py" = 3
 "src/teatree/core/models/review_loop.py" = 3
 "src/teatree/core/models/task.py" = 7
-"src/teatree/core/models/ticket.py" = 13
+"src/teatree/core/models/ticket.py" = 11
 "src/teatree/core/notify.py" = 4
 "src/teatree/core/on_behalf_egress.py" = 2
 "src/teatree/core/on_behalf_gate_recorded.py" = 3
