+++ tests/teatree_loop/test_reviewer_prs_self_authored_1321.py
@@ -31,7 +31,8 @@ from django.test import TestCase
 
 from teatree.core.backend_protocols import PrOpenState, ReviewState
 from teatree.core.models.task import Task
-from teatree.core.models.ticket import Ticket, schedule_external_review
+from teatree.core.models.ticket import Ticket
+from teatree.core.models.ticket_external_review import schedule_external_review
 from teatree.loop.dispatch import dispatch
 from teatree.loop.mechanical import HANDLERS
 from teatree.loop.scanners.reviewer_prs import ReviewerPrsScanner
